Move cx structs to its own file for better readability Make GetType and GetAssignmentElement into a method of cxarg
