--- cx/ast/cxarguments_slice.go
@@ -1,13 +1,0 @@
-package ast
-
-type CXArgumentSlice struct {
-	// Lengths is used if the `CXArgument` defines an array or a
-	// slice. The number of dimensions for the array/slice is
-	// equal to `len(Lengths)`, while the contents of `Lengths`
-	// define the sizes of each dimension. In the case of a slice,
-	// `Lengths` only determines the number of dimensions and the
-	// sizes are all equal to 0 (these 0s are not used for any
-	// computation).
-	Lengths []int
-	IsSlice bool
-}
